Major API change in creating value object classes
Was:
   Person = GS::Value.new(name: String, ...).default(...).create

Now:
   Person = GS::Value.create(name: String, ...) do
     default ...
   end

Or just (if no modification is necessary):
   Person = GS::Value[name: String, ...]

No .create() needed in that last one.

Unfortunately, you can't do "Value[...] do ... end". That's a syntax error.
So I currently have two styles, and I don't like that. One option would
be

   Simple  = Value[...]
   Complex = Value[...].change do
     default ...
     other methods ...
   end

Or of course I could have Value.create(...) for all use cases and get
rid of Value[...] altogether.

Note: this commit also has some improvements to method documentation,
not necessarily related to the code that was changed.