No more GS::Value[...], only GS::Value.create(...)
Only documentation changed in this commit. It seems I made this decision
in code a while ago and forgot about it. There is a syntax issue with []
and do ... end. But also, on reflection, there is a clarity issue as
well. Creating a value object class is a one-time action, so making the
method name clear is beneficial; there's no need for spurious
abbreviation.